requirements.txt
requests>=2.26.0
selenium>=3.141.0
mamba>=0.11.2
expects>=0.9.0
pytest>=6.2.4
webdriver-manager>=3.4.2
pytest-html>=3.1.1
pytest-testconfig>=0.2.0
allure-pytest>=2.9.43
pytest-failed-screenshot>=1.0.2
